project/ajudicator/requirements.txt
Flask==3.0.0
schedule==1.2.1
mysql-connector-python==8.2.0